# internal/api/openapi.json
{
  "components": {
    "schemas": {
...
      "IdentityResponse": {
        "description": "`GET /api/v1/identity` response: the caller's resolved identity.",
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "groups": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "roles": {
            "description": "Snake_case role names (\"viewer\", \"developer\", \"operator\", \"admin\").",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "subject": {
            "type": "string"
          },
          "projects": {
            "description": "The projects the caller holds scoped grants in, from group-derived\nproject roles and stored assignments alike, sorted by name.\n\nDistinct from `roles`, which lists the caller's *global* roles: a global\noperator may act in every project and this list can name none, so a\nclient that finds it empty should read `roles` before concluding the\ncaller may do nothing.",
            "items": {
              "$ref": "#/components/schemas/ProjectGrant"
            },
            "type": "array"
          }
        },
        "required": [
          "subject",
          "groups",
          "roles",
          "projects"
        ],
        "type": "object"
      },
      "ImageConfig": {
        "description": "The image config blob as the OCI image spec defines it: what a container from this image starts with.",
        "properties": {
          "cmd": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "entrypoint": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "env": {
            "description": "Environment variables, split at the first `=`.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "exposed_ports": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "labels": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            },
            "type": "object"
          },
          "user": {
            "type": "string"
          },
          "working_dir": {
            "type": "string"
          }
        },
        "required": [
          "env",
          "entrypoint",
          "cmd",
          "user",
          "working_dir",
          "exposed_ports",
          "labels"
        ],
        "type": "object"
      },
...
      "ImageHistoryEntry": {
        "description": "One `history` row of the image config \u2014 the instruction that produced a layer (BuildKit/docker record the Dockerfile line in `created_by`), joined to the layer it produced when it produced one. This is the closest an image gets to its Dockerfile.",
        "properties": {
          "comment": {
            "type": [
              "string",
              "null"
            ]
          },
          "created": {
            "description": "RFC 3339 timestamp when recorded.",
            "type": [
              "string",
              "null"
            ]
          },
          "created_by": {
            "type": "string"
          },
          "empty_layer": {
            "description": "True for instructions that changed only metadata (ENV, LABEL, CMD, \u2026).",
            "type": "boolean"
          },
          "layer_digest": {
            "description": "Digest of the layer this step produced; `null` for an empty layer.",
            "type": [
              "string",
              "null"
            ]
          },
          "size_bytes": {
            "format": "int64",
            "minimum": 0,
            "description": "Compressed size of that layer; `null` for an empty layer.",
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "created_by",
          "empty_layer"
        ],
        "type": "object"
      },
      "ImageInspect": {
        "description": "What a catalog image looks like without pulling it (#10): the manifest and config blob read from its registry \u2014 platform, size, the container config (env, entrypoint, user, labels), the per-layer build history and the layers. One shape, shared with any other console that renders it.",
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ImageConfig"
          },
          "digest": {
            "description": "Digest of the manifest that was inspected (the platform manifest when `reference` resolved to an index).",
            "type": "string"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ImageHistoryEntry"
            },
            "type": "array"
          },
          "layers": {
            "items": {
              "$ref": "#/components/schemas/ImageLayer"
            },
            "type": "array"
          },
          "platforms": {
            "description": "Every platform the reference offers (one for a single-platform manifest).",
            "items": {
              "$ref": "#/components/schemas/ImagePlatform"
            },
            "type": "array"
          },
          "reference": {
            "description": "The reference that was inspected, as the catalog entry names it.",
            "type": "string"
          },
          "size_bytes": {
            "description": "Sum of the inspected manifest's compressed layer sizes.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "source": {
            "description": "Which producer filled the document: `registry` (Bifrost read the manifest and config itself).",
            "type": "string"
          }
        },
        "required": [
          "reference",
          "digest",
          "platforms",
          "size_bytes",
          "config",
          "history",
          "layers",
          "source"
        ],
        "type": "object"
      },
      "ImageLayer": {
        "properties": {
          "digest": {
            "type": "string"
          },
          "media_type": {
            "type": "string"
          },
          "size_bytes": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "digest",
          "size_bytes",
          "media_type"
        ],
        "type": "object"
      },
      "ImagePlatform": {
        "properties": {
          "architecture": {
            "type": "string"
          },
          "os": {
            "type": "string"
          },
          "variant": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "os",
          "architecture"
        ],
        "type": "object"
      },
...
      "ProjectGrant": {
        "description": "One project the caller holds a scoped grant in, and the roles held\nthere. A client that has to name a project \u2014 the JupyterLab extension\nstarting a cluster, say \u2014 reads this rather than guessing a default.",
        "properties": {
          "name": {
            "description": "The project's name, as `ClusterSpec.project` takes it.",
            "type": "string"
          },
          "roles": {
            "description": "Snake_case role names held *in this project*: \"viewer\", \"developer\",\n\"operator\", \"admin\". A caller needs \"operator\" (or \"admin\") to create\nor stop a cluster there.",
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "name",
          "roles"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "FOSS control plane for Ray clusters. All endpoints take a Bearer JWT (OIDC). Note: the Ray Jobs API is NOT documented here \u2014 it is served by hostname, each registered cluster's hostname proxying Ray's own /api/jobs/ surface through the federating gateway. This spec covers Bifrost's own control-plane routes.",
    "license": {
      "name": "Apache-2.0"
    },
    "title": "Bifrost",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/identity": {
      "get": {
        "operationId": "identity",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            },
            "description": "The caller's identity (the dev identity when auth is disabled)"
          },
          "401": {
            "description": "No/invalid token (only when auth is configured)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "\"Who am I\": the resolved identity for any authenticated caller. In dev\nmode (no validator AND no local auth) the auth middleware attaches no\nidentity, and this returns the specced dev identity so the\nunauthenticated dev loop renders the full console.",
        "tags": [
          "access"
        ]
      }
    },
...
    "/api/v1/images/{name}/inspect": {
      "get": {
        "operationId": "inspect_image",
        "parameters": [
          {
            "description": "Image catalog name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageInspect"
                }
              }
            },
            "description": "The image as its registry describes it"
          },
          "401": {
            "description": "No/invalid token"
          },
          "403": {
            "description": "Missing Read on cluster"
          },
          "404": {
            "description": "No such image, or not open to the caller's projects"
          },
          "502": {
            "description": "The registry could not be reached or refused the read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "images"
        ]
      }
    },
...
  },
...
}